The Final Version of Async_Await_image_coursel
SO I changed the functionality. The major issue was array as an argument and that was causing an issue that I was unable to resolve. But I changed now I am not passing it as an argument. Now I just stored into another array and used conditions then passed the function into setInterval and after every 3 seconds, the image will change.